chore(pkg): move helper around
