+++ internal/controllers/artifact/plugin_controller.go
@@ -82,7 +82,7 @@ func (r *PluginReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctr
 	}
 
 	// Check if the Plugin instance is for the current node.
-	if ok, err := common.NodeMatchesSelector(ctx, r.Client, r.nodeName, plugin.Spec.Selector); err != nil {
+	if ok, err := controllerhelper.NodeMatchesSelector(ctx, r.Client, r.nodeName, plugin.Spec.Selector); err != nil {
 		return ctrl.Result{}, err
 	} else if !ok {
 		logger.Info("Plugin instance does not match node selector, will remove local resources if any")
