+++ internal/controllers/artifact/rulesfile_controller.go
@@ -81,7 +81,7 @@ func (r *RulesfileReconciler) Reconcile(ctx context.Context, req ctrl.Request) (
 	}
 
 	// Check if the Rulesfile instance is for the current node.
-	if ok, err := common.NodeMatchesSelector(ctx, r.Client, r.nodeName, rulesfile.Spec.Selector); err != nil {
+	if ok, err := controllerhelper.NodeMatchesSelector(ctx, r.Client, r.nodeName, rulesfile.Spec.Selector); err != nil {
 		return ctrl.Result{}, err
 	} else if !ok {
 		logger.Info("Rulesfile instance does not match node selector, will remove local resources if any")
