+++ internal/pkg/controllerhelper/node.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package common
+package controllerhelper
 
 import (
 	"context"
